add types folder to contain type definitions
